refactor blogRouter, blog_api.test.js; add test_helpers, update config.js
