+++ part4/controllers/blogs.js
@@ -1,25 +1,32 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/Blog')
 
-blogRouter.get('/', (request, response) => {
-    Blog
-        .find({})
-        .then(blogs => {
-            response.json(blogs)
-        })
+blogRouter.get('/', async (request, response, next) => {
+    try {
+        const blogs = await Blog.find({})
+        response.json(blogs)
+    } catch(error) {
+        next(error)
+    }
 })
 
-blogRouter.post('/', (request, response, next) => {
+blogRouter.post('/', async (request, response, next) => {
     const blog = new Blog(request.body)
 
-    blog
-        .save()
-        .then(result => {
-            response.status(201).json(result)
-        })
-        .catch(error => {
-            next(error)
-        })
+    try {
+        const savedBlog = await blog.save()
+        response.status(201).json(savedBlog)
+    } catch(error)  {
+        next(error)
+    }
+    // blog
+    //     .save()
+    //     .then(result => {
+    //         response.status(201).json(result)
+    //     })
+    //     .catch(error => {
+    //         next(error)
+    //     })
 })
 
 module.exports = blogRouter
