+++ part4/tests/blog_api.test.js
@@ -2,41 +2,42 @@ const mongo = require('mongoose')
 const supertest = require('supertest')
 const app = require('../app')
 const Blog = require('../models/Blog')
-
-const initalBlogs = [
-    {
-        _id: '5a422a851b54a676234d17f7',
-        title: 'React patterns',
-        author: 'Michael Chan',
-        url: 'https://reactpatterns.com/',
-        likes: 7,
-        __v: 0
-    },
-    {
-        _id: '5a422aa71b54a676234d17f8',
-        title: 'Go To Statement Considered Harmful',
-        author: 'Edsger W. Dijkstra',
-        url: 'http://www.u.arizona.edu/~rubinson/copyright_violations/Go_To_Considered_Harmful.html',
-        likes: 5,
-        __v: 0
-    },
-]
+const helper = require('../utils/test_helpers')
 
 beforeEach(async () => {
     await Blog.deleteMany({})
-    let blogObject = new Blog(initalBlogs[0])
-    await blogObject.save()
-    blogObject = new Blog(initalBlogs[1])
-    await blogObject.save()
+
+    // Promise.all() executes promises in parallel
+    /*
+    const blogObjects = helper.initalBlogs.map(blog => new Blog(blog))
+    const promiseBlogsArray = blogObjects.map(blog => blog.save())
+    await Promise.all(promiseBlogsArray)
+    */
+
+    // for of block can execute promises in order
+    for (let each of helper.initalBlogs) {
+        let blog = new Blog(each)
+        await blog.save()
+    }
 })
 
 const api = supertest(app)
 
-test('blog are returned as json', async () => {
-    await api
-        .get('/api/blogs')
-        .expect(200)
-        .expect('Content-Type', /application\/json/)
+describe('get request', () => {
+    test('blog are returned as json', async () => {
+        await api
+            .get('/api/blogs')
+            .expect(200)
+            .expect('Content-Type', /application\/json/)
+    })
+    test('there are two blogs', async() => {
+        const response = await api.get('/api/blogs')
+        expect(response.body).toHaveLength(helper.initalBlogs.length)
+    })
+    test('the first blog has the title React patterns', async () => {
+        const response = await api.get('/api/blogs')
+        expect(response.body[0].title).toBe('React patterns')
+    })
 })
 
 afterAll(() => {
