+++ part4/utils/config.js
@@ -1,4 +1,3 @@
-const { info } = require('../utils/logger')
 require('dotenv').config()
 
 /* global process*/
@@ -7,7 +6,6 @@ const MONGODB_URI = process.env.NODE_ENV === 'test'
     ? process.env.TEST_MONGODB_URI
     : process.env.MONGODB_URI
 
-info('current mongodb uri is' ,MONGODB_URI)
 module.exports = {
     PORT, MONGODB_URI
 }
